Update return value

+++ server/src/main/java/com/tco/misc/GreatCircleDistance.java
@@ -1,5 +1,5 @@
 package com.tco.misc;
 
 public interface GreatCircleDistance extends GeographicCoordinate {
-    public double between(GeographicCoordinate from, GeographicCoordinate to, double earthRadius);
+    public Long between(GeographicCoordinate from, GeographicCoordinate to, double earthRadius);
 }
